Allow lists and tuples as RPC command values
There are a few commands in JunOS that require multiple tags to be sent,
for example the <request-package-add> call when using a set of update
packages for a Virtual-Chassis:

<rpc>
        <request-package-add>
                <reboot/>
                <set>jinstall-ex-4200-...</set>
                <set>jinstall-ex-4500-...</set>
        </request-package-add>
</rpc>

This was not possible as a command value could only be a string. Now it
can also be a tuple or list and a tag will be generated for every value
of the list.