Address issue with fact gathering stopping when it encounters a problem. (#615)
* Address issue with fact gathering stopping when it encounters
a problem. Cleaned up the chassis related fact gathering to
report "UNKNOWN" for model and serial number if it has a problem.

* Updated two unit tests to do with facts gathering via console.
Additional facts are now gathered since we don't stop at the first
exception. The tests had to be updated to account for these additional
facts being gathered.

* Updating test to check for RuntimeError in chassis fact gathering.

* Adding additional tests to improve coverage.

*Fixed an issue where RpcError was not correctly raised
for unknown model or serial number

* Addressing review comments of @vnitinv

* PEP8 cleanup.