Optimize image copying in SW.safe_copy()
The SW.install() method uses the SW.safe_copy() method to copy
a software package to a Junos device if the no_copy argument is False.

This enhancement checks to see if the package is already present on the
Junos device and skips the copy if the package is present and the
file checksum of botht the local and remote packages match.

Also added additional arguments to control the copy and checksum
functionality.