Refactor _exec_rpc() and handle boolean RPC arguments with a value of False.
`_exec_rpc()` returned cryptic errors if an RPC argument was an unexpected
type, or if it had a boolean value of `False`. A boolean with a value of
`False` now results in the argument not being added to the RPC. An argument
with an unexpected type now raises a `TypeError` with an informative message.

Unecessary duplicate code existed to deal with list/tuple argument values vs
string or boolean argument values. Refactored to eliminate the duplication.

Unecessary code to deal with decorator arguments. Consolidated and simplified
the code to deal with decorator arguments.

Additional tests added to cover changes and increase coverage.

PEP8 cleanup.