SW.install() enhanced to install from a URL.
The SW.install() method required either a package or pkg_set argument
to specify the source path (on the local PyEZ host's filesystem) of the
software package. The package was then copied from the PyEZ host to the
remote_path directory of the target device. The software was then
installed from the file which had been copied to the target device.

Junos also supports the ability to specify the software to be installed
using a URL rather than a file on the device's local filesystem.

This enhancement adds that capability to the SW.install() method by
specifying the `url` argument rather than the `package` or `pkg_set`
argument.

i.e.
result = sw.install(url='http://server.example.com/junos-install-mx-x86-64-16.1R4.7.tgz')

URLs can be specified using the formats documented at:
https://www.juniper.net/documentation/en_US/junos/topics/concept/junos-software-formats-filenames-urls.html