Properly handle normalize argument to the open() Device method.
If the normalize argument was specified to the open() method, then
later RPC invocations which attempted to override the normalize value
might not work correctly.

This was because normalizeDecorator() depends on the value of dev._normalize,
but open() was not properly setting the value of dev._normalize.

This change properly sets the value of dev._normalize when the
normalize argument is specified in the call to the open() device method.