SRX Branch cluster fails SW.install(). (#782)
On SRX branch devices configured in an SRX cluster, `SW.install()`
fails because it recognizes the cluster as `'2RE' == True`, and attempts
to upgrade both REs. However, this product doesn't not support the
`re0` and `re1` arguments.

This change avoids attempting to upgrade both REs of the cluster,
even when `all_re = True`.