For MX80 platform, have SW.install() recognize error when package doesn't exist. Fixes #801.
On the MX80 platform. If I attempt to install a package which doesn't exist, then this platform produces:
```
user@mx80> request system software add /var/tmp/nonexistent.tgz | display xml
<rpc-reply xmlns:junos="http://xml.juniper.net/junos/17.3R1/junos">
    <cli>
        <ignore-signals>
            hup
        </ignore-signals>
    </cli>
    <output>
        ERROR: package: /var/tmp/nonexistent.tgz is not found or empty
    </output>
    <cli>
        <banner></banner>
    </cli>
</rpc-reply>
```
Notice that the error message produced is NOT an RPC error, but just a text string in an `<output>` element which
happens to contain the string ERROR. There's also no `<package-result>` element.

There are also other known cases where a missing `<package-result>` element indicates the package was installed
successfully.

This change checks for this specific error message by checking if the string 'ERROR: package:' and the string
'is not found or empty' are present in the `<output>` tag(s).